let game people make polls

+++ jarvis.py
@@ -530,7 +530,7 @@ async def on_raw_reaction_remove(payload):
 async def on_message(message):
     # auto-delete polls
     if message.poll:
-        if not check_is_mod(message.author):
+        if not check_is_mod(message.author) and not check_is_dev(message.author):
             await message.delete()
             return
